+++ eol_scons/tools/gcc.py
@@ -4,41 +4,38 @@
 Implement generic compiling profiles for gcc.
 
 This also adds support for GCC-specific features like address and thread
-sanitizing.  When the gcc tool is in use, the environment provides the
-methods env.SanitizeThread() and env.SanitizeAddress() which add the right
-compile and link flags.  Programs compiled with sanitization can then be
-run and their output 'symbolized' using the AsanFilter wrapper.  For
-example, the following runs the gtest executable and converts any asan
-error output into source file and line numbers.
-
-    env.SanitizeAddress()
-    tests = env.Program('tests', ["test_something.cc"])
-    env.Command('test', tests, env.AsanFilter("${SOURCE.abspath}"))
-
-On Fedora systems, package libasan must be installed for GCC sanitized code
-to link.  The asan_symbolize.py script is included directly because there
-does not appear to be a Fedora package which provides it.
+sanitizing.  When the gcc tool is in use, the environment provides the methods
+env.SanitizeThread() and env.SanitizeAddress() to add the right compile and
+link flags.  On Fedora systems, package libasan must be installed for GCC
+sanitized code to link.  As of at least gcc 5.x, sanitized messages are
+already translated into demangled symbols, so the output does not need to be
+passed through the asan_symbolize.py script and c++filt.  Any calls to
+AsanFilter() can be replaced with just the same command.  The
+SanitizeSupported() method has been removed also, since afaik it was never
+used, and the same can be achieved by just checking for the existence of the
+required method, eg, hasattr(env, 'SanitizeAddress').
 """
 
 import SCons.Tool
 import SCons.Tool.gcc
-import os
-import re
 
 
 def Debug(env):
     env.Append(CCFLAGS=['-g'])
     return env
 
+
 def Warnings(env):
     env.Append(CCFLAGS=['-Wall'])
     if 'NOUNUSED' in env:
-        env.Append (CCFLAGS=['-Wno-unused'])
+        env.Append(CCFLAGS=['-Wno-unused'])
     return env
+
 
 def Optimize(env):
     env.Append(CCFLAGS=['-O2'])
     return env
+
 
 def Profile(env):
     env.Append(CCFLAGS=['-pg'])
@@ -46,11 +43,13 @@ def Profile(env):
     env.Append(SHLINKFLAGS=['-pg'])
     return env
 
+
 def SanitizeAddress(env):
     env.Append(CCFLAGS=['-fsanitize=address', '-fno-omit-frame-pointer'])
     env.Append(LINKFLAGS=['-fsanitize=address'])
     env.Append(SHLINKFLAGS=['-fsanitize=address'])
     return env
+
 
 def SanitizeThread(env):
     env.Append(CCFLAGS=['-fsanitize=thread', '-fno-omit-frame-pointer'])
@@ -58,16 +57,6 @@ def SanitizeThread(env):
     env.Append(SHLINKFLAGS=['-fsanitize=thread'])
     return env
 
-def AsanFilter(env, command):
-    """
-    Wrap a shell command so the output is symbolized with ASAN_FILTER.
-
-    This works by setting pipefail in the shell and then piping the output
-    of the command into the filter command.  It might be better to make
-    this an actual Action implementation which filters the action output in
-    the same way as LogAction in the testing.py tool.
-    """
-    return "set -o pipefail; %s 2>&1 | ${ASAN_FILTER}" % (command)
 
 def generate(env):
     SCons.Tool.gcc.generate(env)
@@ -75,35 +64,8 @@ def generate(env):
     env.AddMethod(Debug)
     env.AddMethod(Warnings)
     env.AddMethod(Profile)
-
-    # There's no harm in always adding these to the construction
-    # environment, whether sanitization will be used or not.  This way
-    # environments can filter output from sanitized executables built in a
-    # different environment.
-    asan = os.path.join(os.path.dirname(__file__), "scripts", 
-                        'asan_symbolize.py')
-    asan = os.path.normpath(asan)
-    env.SetDefault(ASAN_SYMBOLIZE=asan)
-    env.SetDefault(CXXFILT='c++filt')
-    env.SetDefault(ASAN_FILTER="${ASAN_SYMBOLIZE} | ${CXXFILT}")
-
-    # The sanitize options require GCC 4.8+, so if we don't have that,
-    # then provide nop methods.
-
-    version = env.get('CCVERSION')
-    if not version or re.match(r'[1234]\.[1-7]\.', version):
-        env.AddMethod((lambda env: env), "SanitizeAddress")
-        env.AddMethod((lambda env: env), "SanitizeThread")
-        # In theory this doesn't hurt anything, and it's possible someone
-        # may need to symbolize output from something that was not
-        # sanitized, but that can be handled later if ever needed.
-        env.AddMethod((lambda env, command: command), "AsanFilter")
-        env.AddMethod((lambda env: False), "SanitizeSupported")
-    else:
-        env.AddMethod(SanitizeAddress)
-        env.AddMethod(SanitizeThread)
-        env.AddMethod(AsanFilter)
-        env.AddMethod((lambda env: True), "SanitizeSupported")
+    env.AddMethod(SanitizeAddress)
+    env.AddMethod(SanitizeThread)
 
 
 def exists(env):